refactor: api 호출 custom hooks 생성 및 적용
